+++ src/auth0/cache/redis_impl.rs
@@ -1,33 +1,46 @@
 use redis::AsyncCommands;
-use serde::Deserialize;
+use serde::{Deserialize, Serialize};
 
+use super::CacheError;
 use crate::auth0::cache::{self, crypto, Cache};
 use crate::auth0::token::Token;
-use crate::auth0::{Auth0Error, Config};
+
+#[derive(Debug, thiserror::Error)]
+pub enum RedisCacheError {
+    #[error(transparent)]
+    Serde(#[from] serde_json::Error),
+    #[error("redis error: {0}")]
+    Redis(#[from] redis::RedisError),
+    #[error("couldn't decrypt stored token: {0}")]
+    Crypto(#[from] crypto::CryptoError),
+}
+
+impl From<RedisCacheError> for super::CacheError {
+    fn from(val: RedisCacheError) -> Self {
+        CacheError(Box::new(val))
+    }
+}
 
 #[derive(Clone, Debug)]
 pub struct RedisCache {
     client: redis::Client,
     encryption_key: String,
-    caller: String,
-    audience: String,
 }
 
 impl RedisCache {
-    pub async fn new(config_ref: &Config) -> Result<Self, Auth0Error> {
-        let client: redis::Client = redis::Client::open(config_ref.cache_type().redis_connection_url())?;
+    /// Redis connection string(eg. `"redis://{host}:{port}?{ParamKey1}={ParamKey2}"`)
+    pub async fn new(redis_connection_url: String, token_encryption_key: String) -> Result<Self, RedisCacheError> {
+        let client: redis::Client = redis::Client::open(redis_connection_url)?;
         // Ensure connection is fine. Should fail otherwise
         let _ = client.get_multiplexed_async_connection().await?;
 
         Ok(RedisCache {
             client,
-            encryption_key: config_ref.token_encryption_key().to_string(),
-            caller: config_ref.caller().to_string(),
-            audience: config_ref.audience().to_string(),
+            encryption_key: token_encryption_key,
         })
     }
 
-    async fn get<T>(&self, key: &str) -> Result<Option<T>, Auth0Error>
+    async fn get<T>(&self, key: &str) -> Result<Option<T>, RedisCacheError>
     where
         for<'de> T: Deserialize<'de>,
     {
@@ -38,23 +51,30 @@ impl RedisCache {
             .await?
             .map(|value| crypto::decrypt(self.encryption_key.as_str(), value.as_slice()))
             .transpose()
+            .map_err(Into::into)
+    }
+
+    async fn put<T: Serialize>(&self, key: &str, lifetime_in_seconds: u64, v: T) -> Result<(), RedisCacheError> {
+        let mut connection = self.client.get_multiplexed_async_connection().await?;
+
+        let encrypted_value: Vec<u8> = crypto::encrypt(&v, self.encryption_key.as_str())?;
+        let _: () = connection.set_ex(key, encrypted_value, lifetime_in_seconds).await?;
+        Ok(())
     }
 }
 
 #[async_trait::async_trait]
 impl Cache for RedisCache {
-    async fn get_token(&self) -> Result<Option<Token>, Auth0Error> {
-        let key: &str = &cache::token_key(&self.caller, &self.audience);
-        self.get(key).await
+    async fn get_token(&self, client_id: &str, audience: &str) -> Result<Option<Token>, CacheError> {
+        let key: &str = &cache::token_key(client_id, audience);
+        self.get(key).await.map_err(Into::into)
     }
 
-    async fn put_token(&self, value_ref: &Token) -> Result<(), Auth0Error> {
-        let key: &str = &cache::token_key(&self.caller, &self.audience);
-        let mut connection = self.client.get_multiplexed_async_connection().await?;
-        let encrypted_value: Vec<u8> = crypto::encrypt(value_ref, self.encryption_key.as_str())?;
-        let expiration = value_ref.lifetime_in_seconds();
-        let _: () = connection.set_ex(key, encrypted_value, expiration).await?;
-        Ok(())
+    async fn put_token(&self, client_id: &str, audience: &str, value_ref: &Token) -> Result<(), CacheError> {
+        let key: &str = &cache::token_key(client_id, audience);
+        self.put(key, value_ref.lifetime_in_seconds(), value_ref)
+            .await
+            .map_err(Into::into)
     }
 }